third_party/blink/web_tests/external/wpt/mathml/presentation-markup/tables/table-default-styles-001-expected.txt

This is a testharness.js-based test.
[FAIL] Default display value on mtable
  assert_equals: expected "inline-table" but got "table"
[FAIL] Default text-align value on mtd
  assert_equals: expected "center" but got "-webkit-center"
Harness: the test ran to completion.
